yet another jumbo-refactor
quoting_service: merged into quotes
quotesFactory trait: removed
quotesRepository traits: merged into one single trait

credit::persistence: extracted to its own top-level module
swap module: split into a folder
