+++ src/persistence/mod.rs
@@ -1,9 +1,5 @@
 // ----- standard library imports
 // ----- extra library imports
 // ----- local modules
-pub mod admin;
-pub mod error;
-pub mod keys;
-pub mod quotes;
-pub mod web;
+pub mod inmemory;
 // ----- local imports
