+++ src/swap/mod.rs
@@ -1,0 +1,10 @@
+// ----- standard library imports
+// ----- extra library imports
+// ----- local modules
+mod error;
+mod service;
+//pub mod web;
+// ----- local imports
+pub use service::KeysRepository;
+#[cfg(test)]
+pub use service::MockKeysRepository;
