+++ src/swap/web.rs
@@ -1,0 +1,12 @@
+
+// ----- standard library imports
+// ----- extra library imports
+// ----- local imports
+use crate::swap::service::Service
+
+
+pub async fn swap_tokens(
+    State(ctrl): State<ProdQuotingService>,
+pub async fn lookup_quote(
+    Path(id): Path<uuid::Uuid>,
+) -> Result<Json<LookUpQuoteReply>> {
